Move wallet cli commands to top level
