+++ src/commands/burn.ts
@@ -10,29 +10,28 @@ import {
   Transaction,
 } from '@ironfish/sdk'
 import { CliUx, Flags } from '@oclif/core'
-import { IronfishCommand } from '../../command'
-import { IronFlag, RemoteFlags } from '../../flags'
-import { selectAsset } from '../../utils/asset'
-import { connectRpcWallet } from '../../utils/clients'
-import { promptCurrency } from '../../utils/currency'
-import { selectFee } from '../../utils/fees'
-import { watchTransaction } from '../../utils/transaction'
+import { IronfishCommand } from '../command'
+import { IronFlag, RemoteFlags } from '../flags'
+import { selectAsset } from '../utils/asset'
+import { connectRpcWallet } from '../utils/clients'
+import { promptCurrency } from '../utils/currency'
+import { selectFee } from '../utils/fees'
+import { watchTransaction } from '../utils/transaction'
 
-export class Mint extends IronfishCommand {
-  static description = 'Mint tokens and increase supply for a given asset'
+export class Burn extends IronfishCommand {
+  static description = 'Burn tokens and decrease supply for a given asset'
 
   static examples = [
-    '$ ironfish wallet:mint --metadata "see more here" --name mycoin --amount 1000',
-    '$ ironfish wallet:mint --assetId 618c098d8d008c9f78f6155947014901a019d9ec17160dc0f0d1bb1c764b29b4 --amount 1000',
-    '$ ironfish wallet:mint --assetId 618c098d8d008c9f78f6155947014901a019d9ec17160dc0f0d1bb1c764b29b4 --amount 1000 --account otheraccount',
-    '$ ironfish wallet:mint --assetId 618c098d8d008c9f78f6155947014901a019d9ec17160dc0f0d1bb1c764b29b4 --amount 1000 --account otheraccount --fee 0.00000001',
+    '$ ironfish wallet:burn --assetId 618c098d8d008c9f78f6155947014901a019d9ec17160dc0f0d1bb1c764b29b4 --amount 1000',
+    '$ ironfish wallet:burn --assetId 618c098d8d008c9f78f6155947014901a019d9ec17160dc0f0d1bb1c764b29b4 --amount 1000 --account otheraccount',
+    '$ ironfish wallet:burn --assetId 618c098d8d008c9f78f6155947014901a019d9ec17160dc0f0d1bb1c764b29b4 --amount 1000 --account otheraccount --fee 0.00000001',
   ]
 
   static flags = {
     ...RemoteFlags,
     account: Flags.string({
       char: 'f',
-      description: 'The account to mint from',
+      description: 'The account to burn from',
     }),
     fee: IronFlag({
       char: 'o',
@@ -42,23 +41,12 @@ export class Mint extends IronfishCommand {
     }),
     amount: IronFlag({
       char: 'a',
-      description: 'Amount of coins to mint in IRON',
+      description: 'Amount of coins to burn',
       flagName: 'amount',
     }),
     assetId: Flags.string({
       char: 'i',
       description: 'Identifier for the asset',
-      required: false,
-    }),
-    metadata: Flags.string({
-      char: 'm',
-      description: 'Metadata for the asset',
-      required: false,
-    }),
-    name: Flags.string({
-      char: 'n',
-      description: 'Name for the asset',
-      required: false,
     }),
     confirm: Flags.boolean({
       default: false,
@@ -73,7 +61,7 @@ export class Mint extends IronfishCommand {
     rawTransaction: Flags.boolean({
       default: false,
       description:
-        'Return the raw transaction. Used to create a transaction but not post to the network',
+        'Return raw transaction. Use it to create a transaction but not post to the network',
     }),
     expiration: Flags.integer({
       char: 'e',
@@ -91,7 +79,7 @@ export class Mint extends IronfishCommand {
   }
 
   async start(): Promise<void> {
-    const { flags } = await this.parse(Mint)
+    const { flags } = await this.parse(Burn)
     const client = await connectRpcWallet(this.sdk, {
       connectNodeClient: !flags.offline,
     })
@@ -120,51 +108,22 @@ export class Mint extends IronfishCommand {
       account = response.content.account.name
     }
 
-    if (flags.expiration !== undefined && flags.expiration < 0) {
-      this.log('Expiration sequence must be non-negative')
-      this.exit(1)
-    }
-
     let assetId = flags.assetId
-    let metadata = flags.metadata
-    let name = flags.name
 
-    // We can assume the prompt can be skipped if at least one of metadata or
-    // name is provided
-    let isMintingNewAsset = Boolean(name || metadata)
-    if (!assetId && !metadata && !name) {
-      isMintingNewAsset = await CliUx.ux.confirm(
-        'Do you want to create a new asset (Y/N)?',
-      )
-    }
-
-    if (isMintingNewAsset) {
-      if (!name) {
-        name = await CliUx.ux.prompt('Enter the name for the new asset', {
-          required: true,
-        })
-      }
-
-      if (!metadata) {
-        metadata = await CliUx.ux.prompt('Enter metadata for the new asset', {
-          default: '',
-          required: false,
-        })
-      }
-    } else if (!assetId) {
+    if (assetId == null) {
       const asset = await selectAsset(client, account, {
-        action: 'mint',
+        action: 'burn',
         showNativeAsset: false,
-        showNonCreatorAsset: false,
+        showNonCreatorAsset: true,
         showSingleAssetChoice: true,
         confirmations: flags.confirmations,
       })
 
-      if (!asset) {
-        this.error(`You must have an existing asset. Try creating a new one.`)
-      }
+      assetId = asset?.id
+    }
 
-      assetId = asset.id
+    if (assetId == null) {
+      this.error(`You must have a custom asset in order to burn.`)
     }
 
     let amount = flags.amount
@@ -172,20 +131,23 @@ export class Mint extends IronfishCommand {
       amount = await promptCurrency({
         client: client,
         required: true,
-        text: 'Enter the amount',
+        text: 'Enter the amount of the custom asset to burn',
         minimum: 1n,
         logger: this.logger,
+        balance: {
+          account,
+          confirmations: flags.confirmations,
+          assetId,
+        },
       })
     }
 
     const params: CreateTransactionRequest = {
       account,
       outputs: [],
-      mints: [
+      burns: [
         {
           assetId,
-          name,
-          metadata,
           value: CurrencyUtils.encode(amount),
         },
       ],
@@ -195,7 +157,6 @@ export class Mint extends IronfishCommand {
     }
 
     let raw: RawTransaction
-
     if (params.fee === null) {
       raw = await selectFee({
         client,
@@ -219,7 +180,7 @@ export class Mint extends IronfishCommand {
 
     if (
       !flags.confirm &&
-      !(await this.confirm(account, amount, raw.fee, assetId, name, metadata))
+      !(await this.confirm(assetId, amount, raw.fee, account))
     ) {
       this.error('Transaction aborted.')
     }
@@ -235,8 +196,6 @@ export class Mint extends IronfishCommand {
     const transaction = new Transaction(bytes)
 
     CliUx.ux.action.stop()
-
-    const minted = transaction.mints[0]
 
     if (response.content.accepted === false) {
       this.warn(
@@ -254,21 +213,20 @@ export class Mint extends IronfishCommand {
       )
     }
 
-    this.log(
-      `Minted asset ${BufferUtils.toHuman(
-        minted.asset.name(),
-      )} from ${account}`,
+    const assetResponse = await client.wallet.getAsset({
+      account,
+      id: assetId,
+      confirmations: flags.confirmations,
+    })
+    const assetName = BufferUtils.toHuman(
+      Buffer.from(assetResponse.content.name, 'hex'),
     )
-    this.log(`Asset Identifier: ${minted.asset.id().toString('hex')}`)
-    this.log(
-      `Value: ${CurrencyUtils.renderIron(
-        minted.value,
-        true,
-        minted.asset.id().toString('hex'),
-      )}`,
-    )
-    this.log(`Fee: ${CurrencyUtils.renderIron(transaction.fee(), true)}`)
+
+    this.log(`Burned asset ${assetName} from ${account}`)
+    this.log(`Asset Identifier: ${assetId}`)
+    this.log(`Amount: ${CurrencyUtils.renderIron(amount)}`)
     this.log(`Hash: ${transaction.hash().toString('hex')}`)
+    this.log(`Fee: ${CurrencyUtils.renderIron(transaction.fee(), true)}`)
     this.log(
       `\nIf the transaction is mined, it will appear here https://explorer.ironfish.network/transaction/${transaction
         .hash()
@@ -288,20 +246,21 @@ export class Mint extends IronfishCommand {
   }
 
   async confirm(
-    account: string,
+    assetId: string,
     amount: bigint,
     fee: bigint,
-    assetId?: string,
-    name?: string,
-    metadata?: string,
+    account: string,
   ): Promise<boolean> {
-    const nameString = name ? `\nName: ${name}` : ''
-    const metadataString = metadata ? `\nMetadata: ${metadata}` : ''
     this.log(
-      `You are about to mint an asset with the account ${account}:${nameString}${metadataString}`,
+      `You are about to burn: ${CurrencyUtils.renderIron(
+        amount,
+        true,
+        assetId,
+      )} plus a transaction fee of ${CurrencyUtils.renderIron(
+        fee,
+        true,
+      )} with the account ${account}`,
     )
-    this.log(`Amount: ${CurrencyUtils.renderIron(amount, !!assetId, assetId)}`)
-    this.log(`Fee: ${CurrencyUtils.renderIron(fee, true)}`)
 
     return CliUx.ux.confirm('Do you confirm (Y/N)?')
   }
