+++ src/commands/rescan.ts
@@ -4,11 +4,11 @@
 
 import { Meter, TimeUtils } from '@ironfish/sdk'
 import { CliUx, Flags } from '@oclif/core'
-import { IronfishCommand } from '../../command'
-import { RemoteFlags } from '../../flags'
-import { ProgressBar } from '../../types'
-import { hasUserResponseError } from '../../utils'
-import { connectRpcWallet } from '../../utils/clients'
+import { IronfishCommand } from '../command'
+import { RemoteFlags } from '../flags'
+import { ProgressBar } from '../types'
+import { hasUserResponseError } from '../utils'
+import { connectRpcWallet } from '../utils/clients'
 
 export class RescanCommand extends IronfishCommand {
   static description = `Rescan the blockchain for transactions. Clears wallet disk caches before rescanning.`
